+++ src/main.cpp
@@ -1,42 +1,65 @@
 // Project includes
 #include "ArgumentsChecker.h"
-#include "ArgumentsParser.h"
 #include "DatFileUnpacker.h"
 
 // Third party includes
+#include <boost/program_options.hpp>
 
 // stdlib
 #include <iostream>
 
 using namespace DatUnpacker;
 
-void usage()
+void usage(const boost::program_options::options_description& description)
 {
     std::cout << "Unpacker for Fallout 1/2 DAT files" << std::endl;
     std::cout << "v0.0.5 (c) 2012-2022 Falltergeist Developers" << std::endl;
-    std::cout << "Usage: dat-unpacker [arguments]" << std::endl;
     std::cout << "Example: dat-unpacker -f dat1 -s ~/fallout1/master.dat -d ~/unpacked" << std::endl;
+    std::cout << description << std::endl;
     std::cout << std::endl;
-    std::cout << "Arguments:" << std::endl;
-    std::cout << "  --format, -f        Fallout DAT file format version. 'dat1' or 'dat2'. 'dat2' is default" << std::endl;
-    std::cout << "                      Possible values are: 'dat1', 'dat2'." << std::endl;
-    std::cout << "                      (Defaults to 'dat2')" << std::endl;
-    std::cout << "  --quiet, -q         Quite mode. Do not display anything" << std::endl;
-    std::cout << "  --transform, -t     Transform file names to lowercase" << std::endl;
-    std::cout << "  --source, -s        Path to the DAT file" << std::endl;
-    std::cout << "  --destination, -d   Where to extract files" << std::endl;
 }
 
 int main(int argc, char** argv)
 {
-    ArgumentsParser argumentsParser;
-    auto arguments = argumentsParser.parse(argc, argv);
+    namespace po = boost::program_options;
+
+    po::options_description argumentsDescription("Arguments");
+    argumentsDescription.add_options()
+        ("help", "Produce help message")
+        ("format,f", po::value<std::string>()->required()->default_value("dat2"), "Fallout DAT file format version. 'dat1' or 'dat2'. 'dat2' is default")
+        ("quiet,q", po::bool_switch()->default_value(false), "Quite mode. Do not display anything")
+        ("transform,t", po::bool_switch()->default_value(false), "Transform file names to lowercase")
+        ("source,s", po::value<std::string>()->required(), "Path to the DAT file")
+        ("destination,d", po::value<std::string>()->required(), "Where to extract files");
+
+    po::variables_map vm;
+    po::store(po::command_line_parser(argc, argv).options(argumentsDescription).run(), vm);
+
+    if (vm.count("help")) {
+      usage(argumentsDescription);
+      return 1;
+    }
+
+    try {
+      po::notify(vm);
+    } catch (std::exception& e) {
+      std::cout << "Error: " << e.what() << std::endl;
+      usage(argumentsDescription);
+      return 1;
+    }
+
+    Arguments arguments({
+        .quietMode = vm["quiet"].as<bool>(),
+        .transformNames = vm["transform"].as<bool>(),
+        .source = vm["source"].as<std::string>(),
+        .destination = vm["destination"].as<std::string>(),
+        .format = vm["format"].as<std::string>() == "dat2" ? Format::FALLOUT2 : Format::FALLOUT1
+    });
 
     ArgumentsChecker argumentsChecker;
     if (!argumentsChecker.check(arguments)) {
         if (!arguments.quietMode) {
             std::cerr << argumentsChecker.getErrorMessage() << std::endl << std::endl;
-            usage();
         }
         return 1;
     }
@@ -45,7 +68,6 @@ int main(int argc, char** argv)
     if (!datFileUnpacker.unpack(arguments)) {
         if (!arguments.quietMode) {
             std::cerr << datFileUnpacker.getErrorMessage() << std::endl << std::endl;
-            usage();
         }
         return 1;
     }